+++ scripts/types.d.ts
@@ -1,0 +1,17 @@
+export type ComponentsMap = {
+  [name: string]: EventHandler[];
+};
+
+export interface EventHandler {
+  origin: string;
+  name: string;
+  description: string;
+  kind: "function";
+  type: string;
+  handlerArgs: HandlerArg[];
+}
+
+interface HandlerArg {
+  name: string;
+  type: string | undefined;
+}
