+++ src/index.js
@@ -1,5 +1,6 @@
 const fullCorvidTypes = require("../dist/fullCorvidTypes.json");
 const wixModulesNames = require("../dist/wixModules.json");
+const eventHandlers = require("../dist/eventHandlers.json");
 const { TS_CONFIG_PATHS } = require("./constants");
 const dynamicTypings = require("./dynamicTypes");
 
@@ -35,6 +36,7 @@ module.exports = {
     }
   },
   getWixModulesList: () => wixModulesNames,
+  getComponentsEventsHandlers: () => eventHandlers,
   // Methods for Corvid-local (CLI)
   getWidgetTypeDeclarations: dynamicTypings.widget.getRaw,
   getPageElementsTypeDeclarations: dynamicTypings.elementsMap.getRaw,
